fixed minvalue bug for income and outcome and added a confirmation for logoff

+++ src/Components/History/History.js
@@ -66,6 +66,12 @@ export default function History() {
             src={leaveIcon}
             alt="leave"
             onClick={() => {
+              let confirmation = window.confirm(
+                "Tem certeza que deseja fazer logoff?"
+              );
+              if (!confirmation) {
+                return;
+              }
               localStorage.clear("mywallet");
               setUser("");
               navigate("/");
